Factor out WASM formatting helper functions into separate header.
